third_party/WebKit/LayoutTests/external/wpt/web-animations/timing-model/animations/set-the-target-effect-of-an-animation-expected.txt

This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = assert_equals: ready promise is rejected with AbortError expected "AbortError" but got "Error"
FAIL If new effect is null and old effect is not null, we reset the pending tasks and ready promise is rejectedassert_equals: expected "paused" but got "pending"
PASS If animation has a pending pause task, reschedule that task to run as soon as animation is ready.
PASS If animation has a pending play task, reschedule that task to run as soon as animation is ready to play new effect.
FAIL When setting the effect of an animation to the effect of an existing animation, the existing animation's target effect should be set to null.assert_equals: expected "finished" but got "idle"
PASS After setting the target effect of animation to the target effect of an existing animation, the target effect's timing is updated to reflect the current time of the new animation.
Harness: the test ran to completion.
